Convert addon-view info
