+++ .devcontainer/devcontainer.json
@@ -1,0 +1,27 @@
+// See https://aka.ms/vscode-remote/devcontainer.json for format details.
+{
+  "name": "Home Assistant Polymer Dev",
+  "context": "..",
+  "dockerFile": "../Dockerfile",
+  "postCreateCommand": "pip3 install -e .",
+  "appPort": 5050,
+  "runArgs": [
+    "-e",
+    "GIT_EDTIOR='code --wait'",
+    "-e",
+    "NODE_OPTIONS=--max-old-space-size=4096"
+  ],
+  "extensions": [
+    "ms-python.python",
+    "ms-azure-devops.azure-pipelines",
+    "redhat.vscode-yaml"
+  ],
+  "settings": {
+    "python.pythonPath": "/usr/local/bin/python",
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true,
+    "files.trimTrailingWhitespace": true,
+    "editor.rulers": [80],
+    "terminal.integrated.shell.linux": "/bin/bash"
+  }
+}
